Add a uniq parameter to the asset gsp tags
The idea is that if you have a page that is composed of a bunch of modules, you may want each
module to declare it's own asset dependencies. Often times these dependencies will contain a bunch of
overlapping requirements of common libraries and such, which should not be included multiple times.

If the uniq parameter is supplied, keep track of which assets have been included into the page,
or more specifically in the request, and skip writing out duplicate entries. This approach is
completely incompatible with resource bundling, since the bundler can't know where the overlaps are.